Prefer ArrayList over LinkedList
ArrayList uses has less space overhead and better lookup time.
LinkedList is only better when we need to add or remove items
from somewhere besides the end of the list. This never happens
in our code base in practice. Uses of LinkedList should be viewed
with suspicion since it's easy to performance bottlenecks through
use of #get(int index) on a long linked list, as was fixed in
https://github.com/Alluxio/alluxio/pull/6004

This PR changes all but two usages of LinkedList to ArrayList.
All usages for the lists were inspected to make sure we weren't
using any features of the list that would make LinkedList more
appropriate than ArrayList.